refactor(browser): extract undo snapshots and engine replay into project_replay
