+++ crates/vibez-ui/src/app/mod.rs
@@ -138,6 +138,7 @@ mod media;
 mod media_import;
 mod plugins;
 mod project_io;
+mod project_replay;
 mod update;
 mod update_media;
 mod update_remote;
